refactor(types): rename project-type file and update its import
