+++ src/types/index.ts
@@ -3,3 +3,4 @@ export * from './AlertLevel';
 export * from './Tool';
 export * from './SortOption';
 export * from './SidebarItem';
+export * from './Project';
